Moving mqtt code to transport object
